use same tmp directory in all tools, fix ocean postproc on MN4 (still some work to do), fix unitialized vars

+++ timeseries/monitor_ocean.sh
@@ -1,8 +1,8 @@
-#!/bin/ksh
+#!/bin/bash
 
 # Must be called from timeseries.sh, where DATADIR (ie output from hiresclim2) is defined.
 
-set -e
+set -ue
 
 export HERE=`pwd`
 export PYTHONPATH=${HERE}/scripts/barakuda_modules
@@ -39,8 +39,8 @@ while getopts R:y:t:feh option ; do
     esac
 done
 
-
-export TMPDIR_ROOT=$(mktemp -d $SCRATCH/tmp_ecearth3/timeseries_${RUN}_XXXXXX)
+mkdir -p $SCRATCH/tmp_ecearth3/tmp
+export TMPDIR_ROOT=$(mktemp -d $SCRATCH/tmp_ecearth3/tmp/ts_${RUN}_XXXXXX)
 export POST_DIR=$DATADIR
 export DIR_TIME_SERIES=`echo ${DIR_TIME_SERIES} | sed -e "s|<RUN>|${RUN}|g"`
 
@@ -265,8 +265,8 @@ if [ ${IPREPHTML} -eq 0 ]; then
 		    # Creating time vector if first year:
                     if [ ${jv} -eq 0 ]; then
                         rm -f time_${jyear}.nc
-                        ncap2 -3 -h -O -s "time_counter=(time_counter/(24.*3600.)+15.5)/${nbday}" tmp.nc -o tmp0.nc
-			ncap2 -3 -h -O -s "time_counter=time_counter-time_counter(0)+${jyear}+15.5/${nbday}" \
+                        $NCAP -3 -h -O -s "time_counter=(time_counter/(24.*3600.)+15.5)/${nbday}" tmp.nc -o tmp0.nc
+                        $NCAP -3 -h -O -s "time_counter=time_counter-time_counter(0)+${jyear}+15.5/${nbday}" \
                             -s "time_counter@units=\"years\"" tmp0.nc -o tmp2.nc
                         ncks -3 -h -O -v time_counter tmp2.nc -o time_${jyear}.nc
                         rm -f tmp0.nc tmp2.nc
@@ -339,7 +339,8 @@ if [ ${IPREPHTML} -eq 0 ]; then
 
             f_in=$DATADIR/Post_????/${RUN}_${jyear}_moc.nc
 
-            if [ -f ${f_in} ]; then
+            # TODO does not work with ncap
+            if [ -f ${f_in} ] && [ "$NCAP" == "ncap2" ] ; then
 
                 # removing degenerate x dimension:
                 ncwa -3 -O -a x ${f_in} -o moc_tmp.nc
@@ -364,15 +365,15 @@ if [ ${IPREPHTML} -eq 0 ]; then
                     lm1=`expr ${ll} - 1` ; lp1=`expr ${ll} + 1`
 
                     # Only range of latitude and depth we want:
-                    ncap2 -3 -O -h -s "dlat= ((nav_lat >=  ${lm1})&&(nav_lat <  ${lp1}));" -s 'dz= ((depthw < -500)&&(depthw >= -1500));' moc_tmp.nc -o tmp0.nc
-                    ncap2 -3 -O -h -s "x1=dlat*zomsfatl" tmp0.nc -o tmp1.nc
-                    ncap2 -3 -O -h -s "max_amoc_${ll}N=dz*x1"   tmp1.nc -o tmp0.nc
+                    $NCAP -3 -O -h -s "dlat= ((nav_lat >=  ${lm1})&&(nav_lat <  ${lp1}));" -s 'dz= ((depthw < -500)&&(depthw >= -1500));' moc_tmp.nc -o tmp0.nc
+                    $NCAP -3 -O -h -s "x1=dlat*zomsfatl" tmp0.nc -o tmp1.nc
+                    $NCAP -3 -O -h -s "max_amoc_${ll}N=dz*x1"   tmp1.nc -o tmp0.nc
 
                     # Maximum on this remaining y,depthw box:
                     echo "ncwa -O -y max -v max_amoc_${ll}N  -a y,depthw   tmp0.nc -o tmp.nc"
                     ncwa -O -y max -v max_amoc_${ll}N  -a y,depthw   tmp0.nc -o tmp.nc
 
-                    ncap2 -3 -O -h -s "max_amoc_${ll}N@units=\"Sv\"" \
+                    $NCAP -3 -O -h -s "max_amoc_${ll}N@units=\"Sv\"" \
                         -s "max_amoc_${ll}N@long_name=\"Maximum of Atlantic MOC at ${ll}N\""  tmp.nc -o tmp.nc
 
                     rm -f tmp1.nc tmp0.nc
@@ -406,7 +407,8 @@ if [ ${IPREPHTML} -eq 0 ]; then
             cvar="iiceconc"
             f_in=$DATADIR/Post_????/${RUN}_${jyear}_${cvar}.nc
 
-            if [ -f ${f_in} ]; then
+            # TODO does not work with ncap
+            if [ -f ${f_in} ] && [ "$NCAP" == "ncap2" ] ; then
 
                 rm -f tmp*.nc
 
@@ -420,10 +422,10 @@ if [ ${IPREPHTML} -eq 0 ]; then
                 echo "ncwa -3 -O -a t,z tmp.nc mask_no_tz.nc ; rm tmp.nc"
                 ncwa -3 -O -a t,z tmp.nc mask_no_tz.nc ; rm tmp.nc
 
-                ncap2 -3 -A -h -s 'mN= (nav_lat >=  50);' mask_no_tz.nc -o mask_no_tz.nc
-                ncap2 -3 -A -h -s 'mS= (nav_lat <= -45);' mask_no_tz.nc -o mask_no_tz.nc
-                ncap2 -3 -A -h -s "tmasknorth=mN*tmask"   mask_no_tz.nc -o mask_no_tz.nc
-                ncap2 -3 -A -h -s "tmasksouth=mS*tmask"   mask_no_tz.nc -o mask_no_tz.nc
+                $NCAP -3 -A -h -s 'mN= (nav_lat >=  50);' mask_no_tz.nc -o mask_no_tz.nc
+                $NCAP -3 -A -h -s 'mS= (nav_lat <= -45);' mask_no_tz.nc -o mask_no_tz.nc
+                $NCAP -3 -A -h -s "tmasknorth=mN*tmask"   mask_no_tz.nc -o mask_no_tz.nc
+                $NCAP -3 -A -h -s "tmasksouth=mS*tmask"   mask_no_tz.nc -o mask_no_tz.nc
 
                 echo "ncks -3 -O -v tmasknorth,tmasksouth  mask_no_tz.nc   -o tmp0.nc ; rm -f mask_no_tz.nc"
                 ncks -3 -O -v tmasknorth,tmasksouth  mask_no_tz.nc   -o tmp0.nc ; rm -f mask_no_tz.nc
@@ -458,16 +460,16 @@ if [ ${IPREPHTML} -eq 0 ]; then
 
                     cv=${cvar}_${hs}
 
-                    echo "ncap2 -3 -h -O -s "${cv}=${cvar}*tmask${hs}"  tmp0.nc -o tmp_${hs}.nc"
-                    ncap2 -3 -h -O -s "${cv}=${cvar}*tmask${hs}"  tmp0.nc -o tmp_${hs}.nc
+                    echo "$NCAP -3 -h -O -s "${cv}=${cvar}*tmask${hs}"  tmp0.nc -o tmp_${hs}.nc"
+                    $NCAP -3 -h -O -s "${cv}=${cvar}*tmask${hs}"  tmp0.nc -o tmp_${hs}.nc
 
                     echo "ncks -3 -A -v e1t,e2t tmp_e1t_e2t.nc tmp_${hs}.nc"
                     ncks -3 -A -v e1t,e2t tmp_e1t_e2t.nc tmp_${hs}.nc
 
-                    echo "ncap2 -3 -h -O -s "area_ice_${hs}=${cv}*e1t*e2t" tmp_${hs}.nc -o tmp_${hs}.nc"
-                    ncap2 -3 -h -O -s "area_ice_${hs}=${cv}*e1t*e2t" tmp_${hs}.nc -o tmp_${hs}.nc
+                    echo "$NCAP -3 -h -O -s "area_ice_${hs}=${cv}*e1t*e2t" tmp_${hs}.nc -o tmp_${hs}.nc"
+                    $NCAP -3 -h -O -s "area_ice_${hs}=${cv}*e1t*e2t" tmp_${hs}.nc -o tmp_${hs}.nc
 
-                    ncap2 -3 -h -O -s "tot_area_ice_${hs}=area_ice_${hs}.total(\$y,\$x)" \
+                    $NCAP -3 -h -O -s "tot_area_ice_${hs}=area_ice_${hs}.total(\$y,\$x)" \
                         -s "tot_area_ice_${hs}@units=\"m^2\"" tmp_${hs}.nc -o tmp_${hs}.nc
 
                     echo "ncks -3 -A -h -v time_counter time_${jyear}.nc -o tmp_${hs}.nc"
@@ -520,7 +522,7 @@ if [ ${IPREPHTML} -eq 0 ]; then
     echo
 
     #Concatenate new and old files... 
-    if [[ ! -z ${BASE_YEAR_INI} ]] ; then
+    if [[ ! -z ${BASE_YEAR_INI:-} ]] ; then
          echo " Concatenate old and new netcdf files... " 
          ncrcat -h ${OLD_SUPA_FILE} ${SUPA_FILE} ${DIAG_D}/${RUN}_${BASE_YEAR_INI}_${YEAR_END}_time-series_ocean.nc
          rm ${OLD_SUPA_FILE} ${SUPA_FILE}
@@ -574,6 +576,3 @@ if [ ${IPREPHTML} -eq 1 ]; then
     echo; echo
 
 fi # [ ${IPREPHTML} -eq 1 ]
-
-
-rm -rf ${TMPD}
